refactor: EmptyPiece 싱글턴 패턴 적용
