+++ src/main/java/domain/piece/EmptyPiece.java
@@ -1,15 +1,21 @@
 package domain.piece;
 
 import domain.Game;
-import domain.Position;
+import domain.coordinate.Position;
 import domain.Side;
 
 import java.util.List;
 
 public class EmptyPiece extends Piece {
 
-    public EmptyPiece() {
+    private static final EmptyPiece INSTANCE = new EmptyPiece();
+
+    private EmptyPiece() {
         super(Side.NEUTRAL);
+    }
+
+    public static EmptyPiece getInstance() {
+        return INSTANCE;
     }
 
     @Override
